finance_manager/requirements.txt
# Flask Framework Packages
click
Flask
itsdangerous
Jinja2
MarkupSafe
Werkzeug

# Flask Functionality Packages
Flask-Login
Flask-Migrate
Flask-Script
Flask-SQLAlchemy
Flask-WTF
Flask-Admin
WTForms
Flask-Limiter
pyotp
flask-qrcode
cryptography
email_validator
python-dotenv
Flask-Talisman
argon2-cffi

# MySQL Connector
mysql-connector-python
mysqlclient